requirements.txt
absl-py==1.2.0
aiohttp==3.8.1
aiohttp-cors==0.7.0
aiorwlock==1.3.0
aiosignal==1.2.0
alabaster==0.7.12
alembic==1.8.1
antlr4-python3-runtime==4.9.3
anyio==3.6.1
asgiref==3.5.2
asttokens==2.0.8
async-timeout==4.0.2
attrs==22.1.0
autopage==0.5.1
Babel==2.10.3
backcall==0.2.0
backports.functools-lru-cache==1.6.4
beautifulsoup4==4.11.1
black==22.8.0
blessed==1.19.1
blinker==1.4
boto3==1.22.6
botocore==1.25.13
brotlipy==0.7.0
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.0.4
cliff==4.0.0
cloudpickle==2.0.0
cmaes==0.8.2
cmd2==2.4.2
colorama==0.4.5
colorful==0.5.4
colorlog==6.7.0
contourpy==1.0.5
cryptography==37.0.1
cycler==0.11.0
cytoolz==0.12.0
dask==2022.9.1
data==0.4
debugpy==1.6.3
decorator==5.1.1
distlib==0.3.6
docutils==0.19
easydict==1.9
entrypoints==0.4
et-xmlfile==1.0.1
executing==1.1.0
fastapi==0.85.0
fastjsonschema==2.16.2
filelock==3.8.0
fonttools==4.37.3
frozenlist==1.3.1
fsspec==2022.8.2
funcsigs==1.0.2
furo==2022.9.29
future==0.18.2
google-api-core==2.10.2
google-auth==2.11.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.56.4
gpustat==1.0.0
graphviz==0.20.1
greenlet==1.1.3
grpcio==1.43.0
h11==0.14.0
hydra-core==1.2.0
idna==3.4
imagecodecs==2022.9.26
imageio==2.22.0
imagesize==1.4.1
importlib-metadata==4.11.4
iniconfig==1.1.1
ipykernel==6.16.0
ipython==8.4.0
isort==5.10.1
jedi==0.18.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jsonschema==4.16.0
jupyter_client==7.3.5
jupyter_core==4.11.1
kiwisolver==1.4.4
latex==0.7.0
locket==1.0.0
Mako==1.2.3
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.1
matplotlib-inline==0.1.6
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
msgpack==1.0.4
multidict==6.0.2
munkres==1.1.4
mypy-extensions==0.4.3
nbformat==5.7.0
nest-asyncio==1.5.5
networkx==2.8.6
numpy==1.23.4
nvidia-ml-py==11.495.46
oauthlib==3.2.1
omegaconf==2.2.3
opencensus==0.11.0
opencensus-context==0.1.3
openpyxl==3.0.10
optuna==3.0.3
packaging==21.3
pandas==1.5.1
parso==0.8.3
partd==1.3.0
pathspec==0.10.1
pbr==5.10.0
pexpect==4.8.0
pickle5==0.0.11
pickleshare==0.7.5
Pillow==9.2.0
pip==22.2.2
platformdirs==2.5.2
plotly==5.10.0
pluggy==1.0.0
ply==3.11
pockets==0.9.1
prettytable==3.4.1
prometheus-client==0.13.1
prompt-toolkit==3.0.31
protobuf==3.19.6
psutil==5.9.2
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
py-spy==0.3.14
pyarrow==6.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.21
pydantic==1.10.2
Pygments==2.13.0
PyJWT==2.5.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyperclip==1.8.2
PyQt5==5.15.7
PyQt5-sip==12.11.0
pyrsistent==0.18.1
PySocks==1.7.1
pytest==7.1.3
python-dateutil==2.8.2
python-dotenv==0.21.0
pytorch-ignite==0.4.10
pytz==2022.2.1
pyu2f==0.1.5
PyWavelets==1.3.0
PyYAML==6.0
pyzmq==24.0.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.5.2
scikit-image==0.19.3
scikit-learn==1.1.2
scipy==1.8.1
setuptools==65.3.0
shutilwhich==1.1.0
sip==6.6.2
six==1.16.0
smart-open==6.2.0
sniffio==1.3.0
snowballstemmer==2.2.0
soupsieve==2.3.2.post1
Sphinx==5.3.0
sphinx-basic-ng==1.0.0b1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-napoleon==0.7
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
SQLAlchemy==1.4.41
stack-data==0.5.1
starlette==0.20.4
stevedore==4.0.0
tabulate==0.9.0
tempdir==0.7.1
tenacity==8.1.0
tensorboard==2.10.0
tensorboard-data-server==0.6.0
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5.1
threadpoolctl==3.1.0
tifffile==2022.8.12
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==1.12.1
torch-tb-profiler==0.4.0
torchaudio==0.12.1
torchvision==0.13.1
torchviz==0.0.2
tornado==6.2
tqdm==4.64.1
traitlets==5.4.0
typing_extensions==4.3.0
unicodedata2==14.0.0
urllib3==1.26.11
uvicorn==0.16.0
virtualenv==20.16.5
wcwidth==0.2.5
Werkzeug==2.2.2
wheel==0.37.1
yarl==1.7.2
zipp==3.8.1
